да

+++ project/lib/cairo/src/cairo-image-surface-private.h
@@ -40,7 +40,6 @@
 
 #include "cairo-surface-private.h"
 
-#include <stddef.h>
 #include <pixman.h>
 
 CAIRO_BEGIN_DECLS
@@ -72,7 +71,7 @@ struct _cairo_image_surface {
 
     int width;
     int height;
-    ptrdiff_t stride;
+    int stride;
     int depth;
 
     unsigned owns_data : 1;
